requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.20.0
scipy>=1.7.0

# Audio processing
sounddevice>=0.4.5
librosa>=0.9.0
pyaudio>=0.2.11

# Visual processing
opencv-python>=4.5.0
mediapipe>=0.10.0
face-recognition>=1.3.0

# Machine learning
scikit-learn>=1.0.0

# UI and visualization
gradio>=4.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
python-dotenv>=0.19.0
matplotlib>=3.5.0
tqdm>=4.65.0
hydra-core>=1.3.0
